Finish refactoring entry routes
