+++ views/show.ejs
@@ -7,9 +7,7 @@
 <div class="spacer"></div>
 
 <div class="ui text container center aligned main">
-    
-    <!--<h2 class="ui inverted header"></h2>-->
-    
+        
     <div class="ui left aligned inverted container">
         <div class="ui horizontal list">
             <h4 class="ui inverted item header"><%= moment(entry.date).format("D MMM YYYY, h:mm a") %></h4>
@@ -23,50 +21,6 @@
         <h1 class="ui inverted header">Watson data</h1>
         <p class="entrybody"><%- JSON.stringify(entry, null, 4) %></p>
     </div>
-    
-    <!-- <h2 class="ui inverted header">Metadata</h2>
-    <div class="ui divider"></div>
-    <div class="ui doubling stackable grid">
-        <div class="two column row">
-            <div class="chart column" style="position: relative">
-                <div class="ct-chart ct-square" id="chart1"></div>
-                <span class="chart-label"><i class="a2 ui square icon"></i>Order</span>
-                <span class="chart-label"><i class="a1 ui square icon"></i>Chaos</span>
-            </div>
-            <div class="chart column" style="position: relative">
-                <div class="ct-chart ct-square" id="chart2"></div>
-                <span class="chart-label"><i class="b1 ui square icon"></i>Sight</span>
-                <span class="chart-label"><i class="b2 ui square icon"></i>Sound</span>
-                <span class="chart-label"><i class="b3 ui square icon"></i>Smell</span>
-                <span class="chart-label"><i class="b4 ui square icon"></i>Taste</span>
-                <span class="chart-label"><i class="b5 ui square icon"></i>Touch</span>
-            </div>
-        </div>
-        <div class="two column row">
-            <div class="chart column" style="position: relative">
-                <div class="ct-chart ct-square" id="chart3"></div>
-                <span class="chart-label"><i class="c1 ui square icon"></i>Positivity</span>
-                <span class="chart-label"><i class="c2 ui square icon"></i>Anxiety</span>
-                <span class="chart-label"><i class="c3 ui square icon"></i>Sadness</span>
-                <span class="chart-label"><i class="c4 ui square icon"></i>Affection</span>
-                <span class="chart-label"><i class="c5 ui square icon"></i>Aggression</span>
-            </div>
-            <div class="chart column" style="position: relative">
-                <div class="ct-chart ct-square" id="chart4"></div>
-                <span class="chart-label"><i class="d2 ui square icon"></i>Abstract</span>
-                <span class="chart-label"><i class="d1 ui square icon"></i>Concrete</span>
-            </div>
-        </div>
-        <div class="one column row">
-            <div class="chart column" style="position: relative">
-                <div class="ct-chart ct-square" id="chart5"></div>
-                <span class="chart-label"><i class="e2 ui square icon"></i>Work</span>
-                <span class="chart-label"><i class="e1 ui square icon"></i>Social</span>
-            </div>
-        </div>
-    </div>
-    
-    <div class="spacer"></div> -->
     
     <a href="/home" class="ui green button" title="Home"><i class="ui home icon"></i>Home</a>
 </div>
